backend/requirements.txt
Django==3.0.5
djangorestframework==3.12.4
djongo==1.3.4
uuid==1.30
django-cors-headers==3.7.0
requests==2.25.1